multi: allow asset ID to be all zeroes
For v2 addresses we want the address to only show the group key if a
group key is used. The value of the asset ID becomes meaningless for an
address that supports group keys. Unfortunately we can't really _remove_
the value from the address (it's always encoded). But we can set it to
all zeroes so it's more apparent to the user that it's not used.